Finish feature2

+++ src/main/java/ru/eisbrecher/springboot/controller/RESTController.java
@@ -1,5 +1,7 @@
 package ru.eisbrecher.springboot.controller;
 
+
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.eisbrecher.springboot.entity.Fine;
@@ -7,12 +9,14 @@ import ru.eisbrecher.springboot.repository.FineRepository;
 
 import java.util.List;
 
+
 @RestController
 @RequestMapping("/api/v1/fines")
 public class RESTController {
 
     @Autowired
-    FineRepository fineRepository;
+    private FineRepository fineRepository;
+
 
     @GetMapping("/") //получение всех
     public List<Fine> getAllFines() {
@@ -20,7 +24,7 @@ public class RESTController {
     }
 
     @GetMapping("/{fineId}")
-    public Fine getFine(@PathVariable int fineId) {
+    public Fine getFines(@PathVariable int fineId) {
         return fineRepository.getFine(fineId);
     }
 
@@ -36,29 +40,36 @@ public class RESTController {
         return fine;
     }
 
-    @DeleteMapping("/{fineId}")
-    public String removeFine(@RequestParam int fineId) {
-        return fineRepository.removeFine(fineId);
+    @DeleteMapping("/fines/{id}")
+    public String removeFine(@PathVariable int id) {
+        if (fineRepository.getFine(id) == null)
+            return "There is no fine with ID = " + id + " in database";
+        fineRepository.removeFine(id);
+        return "Fine with id = " + id + " was deleted.";
     }
-
     @DeleteMapping("/")
     public String removeFine(@RequestBody Fine fine) {
-        return fineRepository.removeFine(fine);
+        if (fineRepository.getFine(fine.getId()) == null)
+            return "There is no fine with ID = " + fine.getId() + " in database";
+        fineRepository.removeFine(fine);
+        return "Fine with id = " + fine.getId() + " was deleted.";
     }
+
+
+
 
     @PatchMapping("/fines/{fineId}/pay")
     public Fine payFine(@PathVariable int fineId) {
-        Fine fine = getFine(fineId);
+        Fine fine = fineRepository.getFine(fineId);
         fine.setPayed(true);
-        updateFine(fine);
+        fineRepository.updateFine(fine);
         return fine;
     }
-
     @PatchMapping("/fines/{fineId}/court")
     public Fine courtFine(@PathVariable int fineId) {
-        Fine fine = getFine(fineId);
-        fine.setCourt(true);
-        updateFine(fine);
+        Fine fine = fineRepository.getFine(fineId);
+        fine.setPayed(true);
+        fineRepository.updateFine(fine);
         return fine;
     }
